% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_packages.R
\name{available_packages}
\alias{available_packages}
\title{Find available packages}
\usage{
available_packages(source = "CRAN", ...)
}
\arguments{
\item{source}{specify the source for the packages}

\item{...}{further parameters to be passed to the \code{\link{available.packages}} function}
}
\description{
This function is a wrapper for the base \code{available.packages}.
}
\examples{
available_packages()
available_packages(source = "Bioc")
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
