% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{bioc_packages}
\alias{bioc_packages}
\title{Creates the list of packages from Bioconductor to be installed on the server}
\usage{
bioc_packages(packages = NULL)
}
\arguments{
\item{packages}{a character vector specifying packages from Bioconductor repository that you would like to install}
}
\description{
Creates the list of packages from Bioconductor to be installed on the server
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
