% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapr.R
\name{bootstrapr}
\alias{bootstrapr}
\title{Create the bootstrapping script}
\usage{
bootstrapr(write = TRUE, file_name = "bootstrap", file_ext = ".sh", ...)
}
\arguments{
\item{write}{logical indicating whether or not to write the bootstrapping statement to a file}

\item{file_name}{name of file to be written; defaults to "bootstrap"}

\item{file_ext}{character vector of length 1 indicating file extension to use; defaults to ".sh"}
}
\description{
This function creates the bootstrapping script
}
\examples{
bootstrapr()
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
