% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapr_app.R
\name{bootstrapr_app}
\alias{bootstrapr_app}
\title{Run the bootstrapr Shiny app}
\usage{
bootstrapr_app(...)
}
\arguments{
\item{...}{further parameters to be passed to the \code{\link[shiny]{runApp}} function from the Shiny package}
}
\description{
This function launches the \code{bootstrapr} app, which is a Shiny application on a local host. The app features:
\itemize{
\item{interface for selecting packages to add}
}
}
\examples{

\dontrun{
bootstrapr_app()

bootstrapr_app(display.mode = "showcase")
}
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
