% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapr_gui.R
\name{bootstrapr_gui}
\alias{bootstrapr_gui}
\title{Run the bootstrapr GUI}
\usage{
bootstrapr_gui(...)
}
\arguments{
\item{...}{further parameters to be passed to the \code{\link[shiny]{runApp}} function from the Shiny package}
}
\description{
This function launches the \code{bootstrapr} GUI, which is a Shiny application on a local host. The app features:
\itemize{
\item{interface for selecting packages to add}
}
}
\examples{

\dontrun{
bootstrapr_gui()

bootstrapr_gui(display.mode = "showcase")
}
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
