% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dependencies}
\alias{dependencies}
\title{Build dependencies}
\usage{
dependencies(deps = NULL)
}
\arguments{
\item{deps}{system dependencies to be installed, specified as a vector}
}
\description{
This function makes it possible to parameterize tools to be installed
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
