% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{rstudio_server}
\alias{rstudio_server}
\title{RStudio Server}
\usage{
rstudio_server(include = TRUE)
}
\arguments{
\item{include}{logical indicating whether or not you should include the RStudio server}
}
\description{
Helper function to create RStudio Server setup
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
