% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setup}
\alias{setup}
\title{Basic system setup}
\usage{
setup(docker = FALSE)
}
\arguments{
\item{docker}{logical indicating whether or not the bootstrapped script generated should be in DOCKERFILE format}
}
\description{
Helper function to create basic setup for system
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
