% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{shiny_server}
\alias{shiny_server}
\title{Shiny Server}
\usage{
shiny_server(include = TRUE)
}
\arguments{
\item{include}{logical indicating whether or not you should include the Shiny server}
}
\description{
Helper function to create Shiny Server setup
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
