% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapr.R
\name{bootstrapr}
\alias{bootstrapr}
\title{Create the bootstrapping script}
\usage{
bootstrapr(cran = NULL, bioc = NULL, write = FALSE,
  file_name = "bootstrap.sh", docker = FALSE, rstudio_server = TRUE,
  shiny_server = TRUE, cran_repo = getOption("repos")["CRAN"])
}
\arguments{
\item{cran}{character vector indicating the packages to be installed from \url{https://cran.r-project.org/}}

\item{bioc}{character vector indicating the packages to be installed from \url{http://bioconductor.org/}}

\item{write}{logical indicating whether or not to write the bootstrapping statement to a file}

\item{file_name}{name of file to be written; defaults to "bootstrap.sh"}

\item{docker}{logical indicating whether or not the bootstrapped script generated should be in DOCKERFILE format}

\item{rstudio_server}{logical indicating whether or not to include a RStudio Server}

\item{shiny_server}{logical indicating whether or not to include a Shiny Server}

\item{cran_repo}{a character vector of length 1 specifying the CRAN mirror from which you want to install the packages; defaults to user's CRAN mirror global option}
}
\description{
This function creates the bootstrapping script
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}
